Move vesting schedule into its own module
